documents/requirements.txt
azure-ai-textanalytics==5.3.0
azure-ai-formrecognizer==3.3.0
azure-ai-documentintelligence
python-dotenv==0.19.0
azure-storage-blob